requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.65.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5 